[#206] refactor: chatbot 구조 변경
